feat: move UpdateProduct component to shared location for better modularity
